fix: prevent error when fetching reference otus
Consolidate reference otu endpoints into single
view in references.api module.
